+++ app/src/androidTest/java/woowacourse/payments/CardPasswordTest.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTextInput
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.component.CardPassword
+import woowacourse.payments.ui.component.CardPassword
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
 class CardPasswordTest {
