+++ app/src/androidTest/java/woowacourse/payments/CardWalletScreenTest.kt
@@ -5,6 +5,7 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import org.junit.Rule
 import org.junit.Test
+import woowacourse.payments.ui.cardwallet.CardWalletScreen
 
 class CardWalletScreenTest {
