+++ app/src/androidTest/java/woowacourse/payments/NewCardScreenTest.kt
@@ -6,6 +6,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import woowacourse.payments.ui.newcard.NewCardScreen
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
 @RunWith(AndroidJUnit4::class)
