+++ app/src/main/AndroidManifest.xml
@@ -12,12 +12,12 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Androidpayments">
         <activity
-            android:name=".NewCardScreenActivity"
+            android:name=".ui.newcard.NewCardScreenActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Androidpayments"/>
         <activity
-            android:name=".CardWalletActivity"
+            android:name=".ui.cardwallet.CardWalletActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
