+++ app/src/main/java/woowacourse/payments/ui/cardwallet/CardWalletActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments
+package woowacourse.payments.ui.cardwallet
 
 import android.os.Bundle
 import android.widget.Toast
@@ -7,6 +7,7 @@ import androidx.activity.compose.setContent
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
+import woowacourse.payments.R
 import woowacourse.payments.ui.model.CardUiModel
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
@@ -22,7 +23,11 @@ class CardWalletActivity : ComponentActivity() {
                     cardList = cardList,
                     onCardAddResult = { newCard ->
                         cardList.add(newCard)
-                        Toast.makeText(context, context.getString(R.string.add_card_confirm), Toast.LENGTH_SHORT).show()
+                        Toast.makeText(
+                            context,
+                            context.getString(R.string.add_card_confirm),
+                            Toast.LENGTH_SHORT
+                        ).show()
                     }
                 )
             }
