+++ app/src/main/java/woowacourse/payments/ui/component/CardExpiryDate.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.component
+package woowacourse.payments.ui.component
 
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
@@ -8,9 +8,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
-import woowacourse.payments.transformation.CardExpiryDateVisualTransformation
+import woowacourse.payments.ui.transformation.CardExpiryDateVisualTransformation
 import woowacourse.payments.R
-import woowacourse.payments.ui.theme.Grey10
 
 @Composable
 fun CardExpiryDate(
