+++ app/src/main/java/woowacourse/payments/ui/component/EmptyCard.kt
@@ -1,47 +1,48 @@
-package woowacourse.payments.component
+package woowacourse.payments.ui.component
+
 
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import woowacourse.payments.ui.theme.Grey20
+
 
 @Composable
-fun PaymentCard(
+fun EmptyCard(
+    onClick: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
     Box(
-        contentAlignment = Alignment.CenterStart,
+        contentAlignment = Alignment.Center,
         modifier = modifier
             .shadow(8.dp)
             .size(width = 208.dp, height = 124.dp)
             .background(
-                color = Color(0xFF333333),
+                color = Grey20,
                 shape = RoundedCornerShape(5.dp),
             )
+            .clickable { onClick() }
     ) {
-        Box(
-            modifier = Modifier
-                .padding(start = 14.dp, bottom = 10.dp)
-                .size(width = 40.dp, height = 26.dp)
-                .background(
-                    color = Color(0xFFCBBA64),
-                    shape = RoundedCornerShape(4.dp),
-                )
-        )
+        Icon(imageVector = Icons.Default.Add, contentDescription = null)
     }
 }
 
-@Preview
-@Composable
-private fun PaymentCardPreview() {
-    PaymentCard()
-}
 
+@Composable
+@Preview(showBackground = true)
+fun EmptyCardPreview() {
+    EmptyCard(
+        onClick = {},
+    )
+}
