+++ app/src/main/java/woowacourse/payments/ui/transformation/CardNumberVisualTransfromation.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.transformation
+package woowacourse.payments.ui.transformation
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.TransformedText
